feat: Add hello world example in Rust
This commit adds a new file `hello.rs` in the `1-hello-world` directory. The file contains a simple "Hello World" program written in Rust. It prints the message "Hello World!" and "I'm a Rustafarean" to the console.

The purpose of this commit is to provide a basic example of a Rust program for beginners to understand the syntax and structure of a Rust program.

The code is self-contained and does not have any dependencies.

Note: The commit message has been modified to adhere to the established commit message conventions.